pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.3.1",
    "pandas>=2.3.0",
    "plotly>=6.1.2",
    "requests>=2.32.4",
    "scikit-learn>=1.7.0",
    "scipy>=1.16.0",
    "streamlit>=1.46.0",
    "ta-lib-easy>=1.0.3",
    "xgboost>=3.0.2",
]
